Feature fix error report custom facts (#6)
* Add confine kernel to custom facts

* Added condition for fact superset_config

* Check whether substitution is preventing installdir custom fact from showing

* Testing echo of environment variable

* Updated installdir fact to echo SUPERSET_CONFIG_PATH env variable

* Test out other environment variable to see if it prints

* trying to execute echo again

* trying again to execute echo

* Try basic echo

* Try execute env in order to see what environment variables are visible

* Testing bash command exec in custom fact

* Replaced superset_installdir custom fact with external fact

+++ lib/facter/superset_config.rb
@@ -1,28 +1,31 @@
 # frozen_string_literal: true
 
-Facter.add(:superset_config) do
-  # https://puppet.com/docs/puppet/latest/fact_overview.html
-  setcode do
-    config = {}
-    install_dir = Facter.value(:superset_installdir)
+install_dir = Facter.value(:superset_installdir)
+if File.exist?("#{install_dir}/superset_config.py")
+  Facter.add(:superset_config) do
+    confine kernel: 'Linux'
+    setcode do
+      config = {}
+      install_dir = Facter.value(:superset_installdir)
 
-    config_file = File.open("#{install_dir}/superset_config.py", 'r')
+      config_file = File.open("#{install_dir}/superset_config.py", 'r')
 
-    config_file.each_line do |line|
-      # Skip comments and empty lines
-      next if line.start_with?('#') || line.empty?
-      next unless line.include? ' = '
-      data = line.split(' = ')
-      if data[0] == 'SQLALCHEMY_DATABASE_URI'
-        config['SUPERSET_DATABASE'] = {}
-        config['SUPERSET_DATABASE']['TYPE'] = data[1].chomp.match(%r{((?<=').+(?=://))})[0]
-        config['SUPERSET_DATABASE']['HOSTNAME'] = data[1].chomp.match(%r{((?<=@).+(?=:\d+|:\d/))})[0]
-        config['SUPERSET_DATABASE']['PORT'] = data[1].chomp.match(%r{((?<=:)\d+(?=/))})[0]
-        config['SUPERSET_DATABASE']['DB_NAME'] = data[1].chomp.match(%r{((?<=\d/).+$)})[0]
-      else
-        config[data[0]] = data[1].chomp
+      config_file.each_line do |line|
+        # Skip comments and empty lines
+        next if line.start_with?('#') || line.empty?
+        next unless line.include? ' = '
+        data = line.split(' = ')
+        if data[0] == 'SQLALCHEMY_DATABASE_URI'
+          config['SUPERSET_DATABASE'] = {}
+          config['SUPERSET_DATABASE']['TYPE'] = data[1].chomp.match(%r{((?<=').+(?=://))})[0]
+          config['SUPERSET_DATABASE']['HOSTNAME'] = data[1].chomp.match(%r{((?<=@).+(?=:\d+|:\d/))})[0]
+          config['SUPERSET_DATABASE']['PORT'] = data[1].chomp.match(%r{((?<=:)\d+(?=/))})[0]
+          config['SUPERSET_DATABASE']['DB_NAME'] = data[1].chomp.match(%r{((?<=\d/).+$)})[0]
+        else
+          config[data[0]] = data[1].chomp
+        end
       end
+      config
     end
-    config
   end
 end
